Add messaging, notification updates, and recruiter analytics
Implemented messaging between users, including chat based on application ID. Enhanced notifications UI to display messages and alerts dynamically. Added recruiter analytics for tracking total posts and hires. Removed NavbarController to streamline codebase.

+++ src/main/java/com/jobrecruitment/repository/recruiter/RecruiterAnalyticsRepository.java
@@ -3,6 +3,6 @@ package com.jobrecruitment.repository.recruiter;
 import com.jobrecruitment.model.recruiter.RecruiterAnalytics;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RecruiterAnalyticsRepository extends JpaRepository<RecruiterAnalytics, Integer> {
-    RecruiterAnalytics findByRecruiterId(Integer recruiterId);
+public interface RecruiterAnalyticsRepository extends JpaRepository<RecruiterAnalytics, Long> {
+    RecruiterAnalytics findByRecruiterId(Long recruiterId);
 }
